Squashed commit of the following:
commit 6ac4afe9a039f541019e6dec9fb8dfde0b10f658
Merge: ef77e8d 0e8611d
Author: Hubert Futoma <79475962+fhtoma@users.noreply.github.com>
Date:   Sat Jun 19 20:03:09 2021 +0200

    Merge pull request #22 from skni-kod/Turniej

    Turniej

commit 0e8611d1afdcf62310dc3fc919dbb4ed2dc22b0a
Author: fhtoma <futomahubert@gmail.com>
Date:   Sat Jun 19 19:33:16 2021 +0200

    Two Line Update

    Tomek nie umie w console.logi

commit a9acc9df0817e77ed3fdb90cff9eab4fdd3f3a95
Author: Kamil Klecha <urlikodd@gmail.com>
Date:   Sat Jun 19 19:25:48 2021 +0200

    ta galeria nie umie w typescript :KEKW:

commit 4eb954d43f3e06d5ffc55cb6eec4ef00b903ceef
Author: fhtoma <futomahubert@gmail.com>
Date:   Sat Jun 19 19:16:13 2021 +0200

    Gallery Data Base

    Fajna ta galeria

commit 80f2d5990631babe6e428f5f863d4fd518ebb225
Author: fhtoma <futomahubert@gmail.com>
Date:   Sat Jun 19 18:52:04 2021 +0200

    Tournament Players Data Base

    Pobieranie danych o zawodnikach z bazy danych.

commit b5a268635cb2ec7847882a6b2825a4c84c857756
Author: fhtoma <futomahubert@gmail.com>
Date:   Sat Jun 19 18:33:08 2021 +0200

    Error Clearence

    Usunięcie console.logów

commit bdbceda898f52d1d7f596001005a277172619749
Author: fhtoma <futomahubert@gmail.com>
Date:   Sat Jun 19 16:37:36 2021 +0200

    Tournament Title + Files Path Change

    - Reorganizacja folderów
    - Podpięcie TheTournament do bazy
    - Przepisanie galerii na TypeScript

commit 684287971ed80214f8b073e60e7a3dd84041a534
Author: fhtoma <futomahubert@gmail.com>
Date:   Sat Jun 19 15:30:33 2021 +0200

    Fix HomeTabs2

    Usunięcie zbędnych console.log'ów

commit 74a722bfd4cfe618169085b9e625bde57fbb908d
Author: fhtoma <futomahubert@gmail.com>
Date:   Sat Jun 19 15:28:42 2021 +0200

    Fix HomeTabs

    Usunięcie wartości używanej do testów

commit f4ee5bd73106afb032da3c4e9d0a1bdf03933437
Author: fhtoma <futomahubert@gmail.com>
Date:   Sat Jun 19 14:49:45 2021 +0200

    Tournament & Club Data Base

    Podpięcie turniejów i klubów z bazy.

commit ef77e8d9a7d051619682685e24eb4aaded28940b
Merge: 829d0e4 5f92d40
Author: Kamil Klecha <urlikodd@gmail.com>
Date:   Wed Jun 16 15:56:42 2021 +0200

    Merge pull request #21 from skni-kod/login

    Merge

commit 5f92d401e903e89b1046cb4ee37315c562532774
Merge: b63a070 1b0262e
Author: Kamil Klecha <urlikodd@gmail.com>
Date:   Wed Jun 16 15:33:40 2021 +0200

    Merge pull request #20 from skni-kod/Backend

    Proste logowanie

commit b63a07068803313b8d7c7c19b0554ee2467ea36f
Merge: da238a1 829d0e4
Author: Natalia Miarowska <qwertynka15@gmail.com>
Date:   Wed Jun 16 08:36:27 2021 +0200

    Merge branch 'master' into login

commit da238a15f39781a9455db8982f6b12f08275338d
Author: Natalia Miarowska <qwertynka15@gmail.com>
Date:   Tue May 11 16:56:01 2021 +0200

    Update Navbar.vue

commit f9bde296e316014519be4800e7d5a932ab041969
Author: Natalia Miarowska <qwertynka15@gmail.com>
Date:   Tue May 11 16:51:07 2021 +0200

    Login
