+++ public/index.html
@@ -27,3 +27,9 @@
     <!-- built files will be auto injected -->
   </body>
 </html>
+
+<style>
+  #app{
+    background-color: rgb(245, 245, 245);
+  }
+</style>
