+++ src/components/TournamentMatches.vue
@@ -15,6 +15,7 @@
 <script>
 import Vue from 'vue';
 import { Component } from 'vue-property-decorator';
+import axios from '@/axios';
 
 @Component
 export default class Matches extends Vue {
