+++ src/views/TheTournament.vue
@@ -5,20 +5,23 @@
         class="info--text text-h3 text-center font-weight-bold my-8 tytul"
         @click="$router.push('/register')"
       >
-        Wielki Turniej Studenckiego Koła Naukowego Informatyków "KOD"
+        {{ name }}
       </div>
-      
+
       <v-row justify="center" no-gutters class="ma-5">
         <v-col v-for="(mod, i) in modules" :key="i" class="text-center">
+          
           <v-btn
-            @click="$router.replace('/tournament/' + mod.mod)"
+            @click="$router.replace('/tournament/' + id + '/' + mod.mod)"
             color="primary"
             class="rounded text-h5 text-capitalize pa-4"
+            :disabled="$route.path === '/tournament/' + id + '/' + mod.mod"
             >{{ mod.name }}
           </v-btn>
+        
         </v-col>
       </v-row>
-      
+
       <v-card class="py-4 px-6 rounded-lg mb-3">
         <component :is="module"></component>
       </v-card>
@@ -26,12 +29,13 @@
   </v-row>
 </template>
 
-<script lang='ts'>
+<script lang="ts">
 import Vue from 'vue';
-import Gallery from './TournamentGallery.vue';
-import Matches from './TournamentMatches.vue';
-import Players from './TournamentPlayers.vue';
-import Info from './TournamentInfo.vue';
+import Gallery from '@/components/TournamentGallery.vue';
+import Matches from '@/components/TournamentMatches.vue';
+import Players from '@/components/TournamentPlayers.vue';
+import Info from '@/components/TournamentInfo.vue';
+import axios from '@/axios';
 
 import { Component } from 'vue-property-decorator';
 
@@ -44,8 +48,35 @@ import { Component } from 'vue-property-decorator';
   },
 })
 export default class TheTournament extends Vue {
+  created() {
+    this.downloadData();
+  }
+
+  downloadData() {
+    if (this.$store.getters.isAuthenticated) {
+      axios
+        .get('tournament/' + this.$route.params.id + '/', {
+          headers: {
+            Authorization: 'Bearer ' + this.$store.getters.token,
+          },
+        })
+        .then((res2) => {
+          if (res2.status === 200) {
+            let name;
+            const data = res2.data;
+            name = data.name;
+            this.$data.name = name;
+          }
+        })
+        .catch(() => {
+          console.log('Błąd w nazwie turnieju');
+        });
+    }
+  }
+
   data() {
     return {
+      name: '',
       modules: [
         { mod: 'info', name: 'O turnieju' },
         { mod: 'players', name: 'Gracze' },
@@ -54,8 +85,12 @@ export default class TheTournament extends Vue {
       ],
     };
   }
+
   get module() {
     return this.$route.params.module;
+  }
+  get id() {
+    return this.$route.params.id;
   }
 }
 </script>
